+++ swift/core/Sources/A2UICore/Models/SurfaceTheme.swift
@@ -14,3 +14,4 @@
 
 /// A marker protocol representing catalog-defined visual theme parameters.
 public protocol SurfaceTheme: Sendable {}
+